Add caching Enum DescriptionAttribute reader
Extension method for extracting the value of the DescriptionAttribute decorating enumeration values for use in custom stringifying an enum value.
If an enum value is not decorated with the DescriptionAttribute, returns the ToString() value as a fall-back.
Caches the result in a dictionary for performance reasons as the decorating attributes should not be changing at runtime.

REVISIT: Not tested for Flags enums.