Add nvidia-smi outputs to slot model/UI display
The following additional columns are added to the main slot model and as data grid columns to the UI.  These are intended to allow monitoring more information about each slot by making calls to nvidia-smi on each monitored host.
* GPU Fan Speed (%)
* GPU Temperature (C)
* Operating PCIe configuration (i.e. PCIe 4.0 @ x8)
* Maximum system supported PCIe configuration (NOTE: Even if a GPU supports PCIe 4.0, if the motherboard/cables only supports PCIe 3.0, then this will list PCIe 3.0)
* Power State
* Current power draw (Watts)
* Current power limit (Watts)
* Current power limit (%) => This is the percent of the current power limit vs the default power limit
* GPU core clock speed (MHz)
* GPU memory clock speed (MHz)

NOTE: As this is based around nvidia-smi, this will only work for GPUs manufactured by nVidia.