Add project details caching service which queries FAH API
Adds a caching service which loads the projects during start-up, and then loads individual project descriptions (keyed by description id) as they are needed.  Service contains two TryGet methods, one that only gets from what is cached and is intended for use by the UI.  The other will attempt to do a refresh to the API baseed on how recently the last attempt to refresh was made (to limit API requests to a reasonable level).  Initial load of the description will only occur when needed, so if a project has not been encountered by the specific instance of HFM, that particular description will not be loaded.

This is ultimately just to get the project Cause from the API, which is only available as part of the description query.

NOTE 1: This is purely an in-memory cache.  There is not presently any persistence associated with this service.  So on restarting, the FAH API will need to be queried afresh in order to rebuild the cache.
NOTE 2: Initial implementation only hooks this into the main view so user can only view the currently executing work units.  Should be usable as well in the history view, however, need to figure out best way to load the larger set of project IDs so as not to overwhelm the API and trigger "too many requests" bans.  This should be partially mitigated by inclusion of persistence to rebuild the local cache first.