Add SSH configuration to ClientSettings
Includes fields needed to determine if SSH login is enabled and configured, including the user name, port, and private key registered for SSH connection to the remote server.
It is intended that this authentication will be set via functionality to be added in a future commit (automatic setup) or by manually configured settings.
Additional work to the UI needs to be added to populate this information, however, for now manually editing the hfmx config file can be used to provide these values.