Add FAH Project details model (truncated) + deserialization tests
This model is for deserializing a truncated set of properties from the results of a query to the FAH API server on endpoint https://api2.foldingathome.org/project/description/{0}, where the replacement value is the ProjectID value.
The model is truncated to exclude things like icons, thumbnails, URLs, etc, but also contains non-serialization properites for managing timestamps to indicate how recently an update query was made (and the last successful attempt).  These timestamps are to prevent overquerying the API (which unnecessarily overutilzies resources and can potentially cause temporary "bans" for making too many requests).
The primary concern is to obtain the full description and casuse of the project (the cause is what we will ultimately be displaying in the UI).
NOTE: Inclusion of the description field was for a near-future concern that in the historical project views being able to show the description.  This field, however, is a premature concern and doesn't actually need to be included as part of this specific feature implementation (and can be eliminated if needed w/o concequence).
The reason for truncating was to exclude things that would not be used to reduce memory usage by a small amount.  It should be simple/easy to add in additional fields from the response if/when they are needed in the future.
This model is current as of 2023-11-04 and includes a truncated set of fields as mentioned.