Add project summary caching service (no tests)
Adds a caching service which loads/updates the projects summary cache.  Service contains two methods, one that only gets from what is cached and is intended for use by the UI.  The other will attempt to do a refresh to the API baseed on how recently the last attempt to refresh was made (to limit API requests to a reasonable level).

NOTE: This is purely an in-memory cache.  There is not presently any persistence associated with this service (to be implemented when persistence later refactor is completed and merged).  So on restarting, the FAH API will need to be queried afresh in order to rebuild the cache.